# docs/api/openapi.yaml
openapi: 3.0.3
info:
  title: 校园失物招领系统 API
  version: 0.2.0
  description: 终审整改契约：与Controller method+path及实际HTTP行为对应；所有JSON成功为200且含ApiEnvelope，文件下载例外。时间UTC带Z；ID为int64数值。动态文件权限、受理授权见conventions。运行期springdoc用于交叉检查，不能替代此处业务契约。
servers:
- url: /api/v1
  description: 统一前缀
...
security:
- bearerAuth: []
paths:
...
  /posts:
    get:
      tags:
      - M3-Post
      summary: 公开列表（FR-POST-03）
      security: []
      parameters:
      - name: page
        in: query
        schema:
          type: integer
          default: 1
          maximum: 100000
      - name: pageSize
        in: query
        schema:
          type: integer
          default: 20
          maximum: 100
      - name: type
        in: query
        schema:
          type: string
          enum:
          - LOST
          - FOUND
      - name: category
        in: query
        schema:
          type: string
      - name: campus
        in: query
        schema:
          type: string
      responses:
        '200':
          description: OK；统一响应封装
          content:
            application/json:
              schema:
                allOf:
                - $ref: '#/components/schemas/ApiEnvelope'
                - type: object
                  properties:
                    data:
                      type: object
                      required:
                      - items
                      - total
                      - page
                      - pageSize
                      properties:
                        items:
                          type: array
                          items:
                            $ref: '#/components/schemas/PostSummary'
                        total:
                          type: integer
                          format: int64
                        page:
                          type: integer
                          format: int32
                        pageSize:
                          type: integer
                          format: int32
        '400':
          description: 参数格式或业务校验失败
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiEnvelope'
        '403':
          description: 角色或账户受限
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiEnvelope'
        '404':
          description: 不存在或无权访问
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiEnvelope'
        '409':
          description: 状态/唯一约束冲突
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiEnvelope'
    post:
      tags:
      - M3-Post
      summary: 发布寻物/招领（FR-POST-01/02）
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreatePostRequest'
      responses:
        '400':
          description: INVALID_ARGUMENT
        '200':
          description: OK；统一响应封装
          content:
            application/json:
              schema:
                allOf:
                - $ref: '#/components/schemas/ApiEnvelope'
                - type: object
                  properties:
                    data:
                      $ref: '#/components/schemas/PostDetail'
        '401':
          description: 未登录或会话失效
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiEnvelope'
        '403':
          description: 角色或账户受限
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiEnvelope'
        '404':
          description: 不存在或无权访问
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiEnvelope'
        '409':
          description: 状态/唯一约束冲突
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiEnvelope'
...
  /posts/{postId}:
    get:
      tags:
      - M3-Post
      summary: 详情（FR-POST-03）
      security: []
      parameters:
      - name: postId
        in: path
        required: true
        schema:
          type: integer
          format: int64
      responses:
        '200':
          description: OK；统一响应封装
          content:
            application/json:
              schema:
                allOf:
                - $ref: '#/components/schemas/ApiEnvelope'
                - type: object
                  properties:
                    data:
                      $ref: '#/components/schemas/PostDetail'
        '404':
          description: NOT_FOUND
        '400':
          description: 参数格式或业务校验失败
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiEnvelope'
        '403':
          description: 角色或账户受限
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiEnvelope'
        '409':
          description: 状态/唯一约束冲突
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiEnvelope'
    patch:
      tags:
      - M3-Post
      summary: 本人编辑（FR-POST-04，核心字段可能被锁）
      parameters:
      - name: postId
        in: path
        required: true
        schema:
          type: integer
          format: int64
      responses:
        '200':
          description: OK；统一响应封装
          content:
            application/json:
              schema:
                allOf:
                - $ref: '#/components/schemas/ApiEnvelope'
                - type: object
                  properties:
                    data:
                      $ref: '#/components/schemas/PostDetail'
        '409':
          description: POST_EDIT_LOCKED
        '400':
          description: 参数格式或业务校验失败
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiEnvelope'
        '401':
          description: 未登录或会话失效
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiEnvelope'
        '403':
          description: 角色或账户受限
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiEnvelope'
        '404':
          description: 不存在或无权访问
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiEnvelope'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/UpdatePostRequest'
...
  /admin/posts:
    get:
      tags:
      - M2-Admin
      summary: 治理检索发布(FR-ADMIN-01)
      responses:
        '200':
          description: OK；统一响应封装
          content:
            application/json:
              schema:
                allOf:
                - $ref: '#/components/schemas/ApiEnvelope'
                - type: object
                  properties:
                    data:
                      type: object
                      required:
                      - items
                      - total
                      - page
                      - pageSize
                      properties:
                        items:
                          type: array
                          items:
                            $ref: '#/components/schemas/Post'
                        total:
                          type: integer
                          format: int64
                        page:
                          type: integer
                          format: int32
                        pageSize:
                          type: integer
                          format: int32
        '400':
          description: 参数格式或业务校验失败
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiEnvelope'
        '401':
          description: 未登录或会话失效
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiEnvelope'
        '403':
          description: 角色或账户受限
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiEnvelope'
        '404':
          description: 不存在或无权访问
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiEnvelope'
        '409':
          description: 状态/唯一约束冲突
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiEnvelope'
      parameters:
      - name: page
        in: query
        schema:
          type: integer
          default: 1
          maximum: 100000
      - name: pageSize
        in: query
        schema:
          type: integer
          default: 20
          maximum: 100
      - name: status
        in: query
        schema:
          type: string
      - name: type
        in: query
        schema:
          type: string
  /admin/posts/{postId}/remove:
    post:
      tags:
      - M2-Admin
      summary: 下架（FR-ADMIN-01）
      parameters:
      - name: postId
        in: path
        required: true
        schema:
          type: integer
          format: int64
      responses:
        '200':
          description: OK；统一响应封装
          content:
            application/json:
              schema:
                allOf:
                - $ref: '#/components/schemas/ApiEnvelope'
                - type: object
                  properties:
                    data:
                      type: object
                      nullable: true
                      description: 该操作data为null
        '400':
          description: 参数格式或业务校验失败
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiEnvelope'
        '401':
          description: 未登录或会话失效
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiEnvelope'
        '403':
          description: 角色或账户受限
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiEnvelope'
        '404':
          description: 不存在或无权访问
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiEnvelope'
        '409':
          description: 状态/唯一约束冲突
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiEnvelope'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ReasonRequest'
  /admin/posts/{postId}/restore:
    post:
      tags:
      - M2-Admin
      summary: 按规则恢复（FR-ADMIN-01）
      parameters:
      - name: postId
        in: path
        required: true
        schema:
          type: integer
          format: int64
      responses:
        '200':
          description: OK；统一响应封装
          content:
            application/json:
              schema:
                allOf:
                - $ref: '#/components/schemas/ApiEnvelope'
                - type: object
                  properties:
                    data:
                      type: object
                      nullable: true
                      description: 该操作data为null
        '400':
          description: 参数格式或业务校验失败
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiEnvelope'
        '401':
          description: 未登录或会话失效
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiEnvelope'
        '403':
          description: 角色或账户受限
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiEnvelope'
        '404':
          description: 不存在或无权访问
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiEnvelope'
        '409':
          description: 状态/唯一约束冲突
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiEnvelope'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ReasonRequest'
  /admin/users:
    get:
      tags:
      - M2-Admin
      summary: 治理检索用户(FR-ADMIN-02)
      responses:
        '200':
          description: OK；统一响应封装
          content:
            application/json:
              schema:
                allOf:
                - $ref: '#/components/schemas/ApiEnvelope'
                - type: object
                  properties:
                    data:
                      type: object
                      required:
                      - items
                      - total
                      - page
                      - pageSize
                      properties:
                        items:
                          type: array
                          items:
                            $ref: '#/components/schemas/User'
                        total:
                          type: integer
                          format: int64
                        page:
                          type: integer
                          format: int32
                        pageSize:
                          type: integer
                          format: int32
        '400':
          description: 参数格式或业务校验失败
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiEnvelope'
        '401':
          description: 未登录或会话失效
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiEnvelope'
        '403':
          description: 角色或账户受限
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiEnvelope'
        '404':
          description: 不存在或无权访问
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiEnvelope'
        '409':
          description: 状态/唯一约束冲突
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiEnvelope'
      parameters:
      - name: page
        in: query
        schema:
          type: integer
          default: 1
          maximum: 100000
      - name: pageSize
        in: query
        schema:
          type: integer
          default: 20
          maximum: 100
      - name: keyword
        in: query
        schema:
          type: string
  /admin/users/{userId}/restrictions:
    post:
      tags:
      - M2-Admin
      summary: 限制用户（FR-ADMIN-02）
      parameters:
      - name: userId
        in: path
        required: true
        schema:
          type: integer
          format: int64
      responses:
        '200':
          description: OK；统一响应封装
          content:
            application/json:
              schema:
                allOf:
                - $ref: '#/components/schemas/ApiEnvelope'
                - type: object
                  properties:
                    data:
                      type: object
                      nullable: true
                      description: 该操作data为null
        '400':
          description: 参数格式或业务校验失败
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiEnvelope'
        '401':
          description: 未登录或会话失效
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiEnvelope'
        '403':
          description: 角色或账户受限
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiEnvelope'
        '404':
          description: 不存在或无权访问
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiEnvelope'
        '409':
          description: 状态/唯一约束冲突
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiEnvelope'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ReasonRequest'
  /admin/users/{userId}/unrestrict:
    post:
      tags:
      - M2-Admin
      summary: 解除限制（FR-ADMIN-02）
      parameters:
      - name: userId
        in: path
        required: true
        schema:
          type: integer
          format: int64
      responses:
        '200':
          description: OK；统一响应封装
          content:
            application/json:
              schema:
                allOf:
                - $ref: '#/components/schemas/ApiEnvelope'
                - type: object
                  properties:
                    data:
                      type: object
                      nullable: true
                      description: 该操作data为null
        '400':
          description: 参数格式或业务校验失败
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiEnvelope'
        '401':
          description: 未登录或会话失效
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiEnvelope'
        '403':
          description: 角色或账户受限
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiEnvelope'
        '404':
          description: 不存在或无权访问
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiEnvelope'
        '409':
          description: 状态/唯一约束冲突
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiEnvelope'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ReasonRequest'
...
  /admin/posts/{postId}:
    get:
      tags:
      - M2-Admin
      summary: 发布公开详情、图片与治理历史
      responses:
        '200':
          description: OK；统一响应封装
          content:
            application/json:
              schema:
                allOf:
                - $ref: '#/components/schemas/ApiEnvelope'
                - type: object
                  properties:
                    data:
                      $ref: '#/components/schemas/PostGovernanceDetail'
        '400':
          description: 参数格式或业务校验失败
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiEnvelope'
        '401':
          description: 未登录或会话失效
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiEnvelope'
        '403':
          description: 角色或账户受限
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiEnvelope'
        '404':
          description: 不存在或无权访问
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiEnvelope'
        '409':
          description: 状态/唯一约束冲突
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiEnvelope'
      parameters:
      - name: postId
        in: path
        required: true
        schema:
          type: integer
          format: int64
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  schemas:
    ApiEnvelope:
      type: object
      properties:
        code:
          type: string
          example: OK
        message:
          type: string
        data: {}
        requestId:
          type: integer
          nullable: true
          format: int64
...
    PostGovernanceDetail:
      type: object
      properties:
        post:
          $ref: '#/components/schemas/Post'
        imageFileIds:
          type: array
          items:
            type: integer
            format: int64
        history:
          type: array
          items:
            $ref: '#/components/schemas/ModerationAction'
...
    User:
      type: object
      properties:
        id:
          type: integer
          format: int64
        nickname:
          type: string
        campus:
          type: string
        avatarFileId:
          type: integer
          format: int64
        status:
          type: string
        campusVerificationStatus:
          type: string
        createdAt:
          type: string
          format: date-time
          description: UTC ISO 8601，响应带Z
        updatedAt:
          type: string
          format: date-time
          description: UTC ISO 8601，响应带Z
    CreatePostRequest:
      type: object
      properties:
        type:
          type: string
          enum:
          - LOST
          - FOUND
        title:
          type: string
          maxLength: 128
          minLength: 1
        category:
          type: string
          maxLength: 48
          minLength: 1
        publicDescription:
          type: string
          maxLength: 2000
          minLength: 1
        campus:
          type: string
          maxLength: 64
          nullable: true
        eventLocation:
          type: string
          maxLength: 128
          nullable: true
        eventTime:
          type: string
          format: date-time
          description: UTC ISO 8601，响应带Z
          nullable: true
        imageFileIds:
          type: array
          items:
            type: integer
            format: int64
      required:
      - type
      - title
      - category
      - publicDescription
    UpdatePostRequest:
      type: object
      properties:
        title:
          type: string
          maxLength: 128
          nullable: true
        category:
          type: string
          maxLength: 48
          nullable: true
        publicDescription:
          type: string
          maxLength: 2000
          nullable: true
        campus:
          type: string
          maxLength: 64
          nullable: true
        eventLocation:
          type: string
          maxLength: 128
          nullable: true
        eventTime:
          type: string
          format: date-time
          description: UTC ISO 8601，响应带Z
          nullable: true
        imageFileIds:
          type: array
          items:
            type: integer
            format: int64
    PostSummary:
      type: object
      properties:
        id:
          type: integer
          format: int64
          nullable: true
        type:
          type: string
          nullable: true
        title:
          type: string
          nullable: true
        category:
          type: string
          nullable: true
        campus:
          type: string
          nullable: true
        eventLocation:
          type: string
          nullable: true
        eventTime:
          type: string
          format: date-time
          description: UTC ISO 8601，响应带Z
          nullable: true
        publishedAt:
          type: string
          format: date-time
          description: UTC ISO 8601，响应带Z
          nullable: true
        status:
          type: string
          nullable: true
        imageFileIds:
          type: array
          items:
            type: integer
            format: int64
    PostDetail:
      type: object
      properties:
        id:
          type: integer
          format: int64
          nullable: true
        type:
          type: string
          nullable: true
        title:
          type: string
          nullable: true
        category:
          type: string
          nullable: true
        publicDescription:
          type: string
          nullable: true
        campus:
          type: string
          nullable: true
        eventLocation:
          type: string
          nullable: true
        eventTime:
          type: string
          format: date-time
          description: UTC ISO 8601，响应带Z
          nullable: true
        publishedAt:
          type: string
          format: date-time
          description: UTC ISO 8601，响应带Z
          nullable: true
        status:
          type: string
          nullable: true
        version:
          type: integer
          format: int32
          nullable: true
        publisherId:
          type: integer
          format: int64
          nullable: true
        publisherNickname:
          type: string
          nullable: true
        mine:
          type: boolean
        imageFileIds:
          type: array
          items:
            type: integer
            format: int64
    Post:
      type: object
      properties:
        id:
          type: integer
          format: int64
        publisherId:
          type: integer
          format: int64
        type:
          type: string
        title:
          type: string
        category:
          type: string
        categoryCode:
          type: string
        publicDescription:
          type: string
        campus:
          type: string
        eventLocation:
          type: string
        eventTime:
          type: string
          format: date-time
          description: UTC ISO 8601，响应带Z
        publishedAt:
          type: string
          format: date-time
          description: UTC ISO 8601，响应带Z
        status:
          type: string
        version:
          type: integer
          format: int32
        resolvedByClaimId:
          type: integer
          format: int64
        closedAt:
          type: string
          format: date-time
          description: UTC ISO 8601，响应带Z
        createdAt:
          type: string
          format: date-time
          description: UTC ISO 8601，响应带Z
        updatedAt:
          type: string
          format: date-time
          description: UTC ISO 8601，响应带Z
...
    ReasonRequest:
      type: object
      properties:
        reason:
          type: string
          maxLength: 255
          minLength: 1
      required:
      - reason
    ModerationAction:
      type: object
      properties:
        id:
          type: integer
          format: int64
        adminId:
          type: integer
          format: int64
        targetType:
          type: string
        targetId:
          type: integer
          format: int64
        action:
          type: string
        reason:
          type: string
        beforeState:
          type: string
        afterState:
          type: string
        createdAt:
          type: string
          format: date-time
          description: UTC ISO 8601，响应带Z
...